ai_end_refactor/pyproject.toml
[project]
name = "careerpath"
version = "0.1.0"
description = "通用 AI Agent 文档检索后端"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.0",
    "openai>=2.17.0",
    "pyyaml>=6.0.3",
    "python-multipart>=0.0.17",
    "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]
di = [
    "dependency-injector>=4.48.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--color=yes"
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/**"]
